fix(ci): drop dead RunTicks helper, fix int→float promotion
GCC -Werror=unused-function and -Werror=conversion still tripped on
the CMake CI Linux Debug job:

- test_witness_replication.cpp had a RunTicks helper that was never
  called (the inline comment even noted "use AoIMap to assert
  lod_next_update_tick instead — see below tests"). Deleted it.
- test_witness_enter_pending_uaf.cpp built a Vector3 with `1.f + i *
  0.5f` where i is int. The implicit int→float promotion is exactly
  what -Wconversion warns about; static_cast<float>(i) clears it.

+++ tests/unit/test_witness_enter_pending_uaf.cpp
@@ -129,7 +129,7 @@ TEST_F(WitnessEnterPendingUafTest, SendCallbackDestroysAnotherPeer) {
   std::vector<EntityID> peer_ids;
   for (int i = 0; i < 8; ++i) {
     const EntityID id = 200 + i;
-    MakeEntity(id, {1.f + i * 0.5f, 0, 0.5f});
+    MakeEntity(id, {1.f + static_cast<float>(i) * 0.5f, 0, 0.5f});
     peer_ids.push_back(id);
   }
